// store-backend/src/rbac/enums/permission.enum.ts
export enum Permission {
  // --- مدیریت محصولات (Products) ---
  PRODUCT_CREATE = 'PRODUCT_CREATE',
  PRODUCT_READ = 'PRODUCT_READ',
  PRODUCT_UPDATE = 'PRODUCT_UPDATE',
  PRODUCT_DELETE = 'PRODUCT_DELETE',
  PRODUCT_INVENTORY_MANAGE = 'PRODUCT_INVENTORY_MANAGE',

  // --- دسته‌بندی‌ها و برندها (Categories & Brands) ---
  CATEGORY_CREATE = 'CATEGORY_CREATE',
  CATEGORY_READ = 'CATEGORY_READ',
  CATEGORY_UPDATE = 'CATEGORY_UPDATE',
  CATEGORY_DELETE = 'CATEGORY_DELETE',

  // --- سفارشات (Orders) ---
  ORDER_CREATE = 'ORDER_CREATE', // ثبت سفارش توسط کاربر
  ORDER_READ_OWN = 'ORDER_READ_OWN', // دیدن سفارش‌های خود کاربر
  ORDER_READ_ALL = 'ORDER_READ_ALL',
  ORDER_UPDATE_STATUS = 'ORDER_UPDATE_STATUS',
  ORDER_CANCEL = 'ORDER_CANCEL',

  COUPON_CREATE = 'COUPON_CREATE',
  COUPON_READ = 'COUPON_READ',
  COUPON_UPDATE = 'COUPON_UPDATE',
  COUPON_DELETE = 'COUPON_DELETE',
  COUPON_APPLY = 'COUPON_APPLY',

  // --- مدیریت کاربران (Users Management) ---
  USER_READ_ALL = 'USER_READ_ALL',
  USER_UPDATE_ROLE = 'USER_UPDATE_ROLE',
  USER_BLOCK = 'USER_BLOCK',

  REVIEW_CREATE = 'REVIEW_CREATE',
  REVIEW_READ = 'REVIEW_READ',
  REVIEW_MODERATE = 'REVIEW_MODERATE',
  REVIEW_DELETE = 'REVIEW_DELETE',

  DASHBOARD_READ = 'DASHBOARD_READ',
  REPORTS_EXPORT = 'REPORTS_EXPORT',
}